% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_datasets.R
\name{calculate_taxes}
\alias{calculate_taxes}
\title{Calculate Taxes}
\usage{
calculate_taxes(family_types)
}
\arguments{
\item{family_types}{A tibble containing family compositions, sizes, and monthly incomes.
Use \code{create_benefits_table()} to calculate family types}
}
\value{
A tibble with calculated federal and state taxes, including details of deductions and credits applied.
}
\description{
Computes federal and state taxes for a household based on their income and family composition.
The calculation includes adjustments for deductions, credits, and other tax-related considerations.
}
