% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_cleaning_functions.R
\name{check_parameters}
\alias{check_parameters}
\title{Validate Parameter Values}
\usage{
check_parameters(actual_data, test_data, test_name)
}
\arguments{
\item{actual_data}{A vector of values to be validated.}

\item{test_data}{A vector of valid reference values.}

\item{test_name}{Character. A name used to describe the type of data being tested (e.g., "value_name").}
}
\value{
Throws an error if validation fails; otherwise, the function returns nothing.
}
\description{
Checks if all elements in a data vector are valid based on a reference vector.
If any element is invalid, the function throws an error with a descriptive message.
}
