% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/federal_poverty_guidelines.R
\name{get_single_poverty_guidelines}
\alias{get_single_poverty_guidelines}
\title{Fetch Federal Poverty Guidelines for single household size}
\usage{
get_single_poverty_guidelines(year, state, household_size, by_month = FALSE)
}
\arguments{
\item{year}{An integer representing the year. Valid years are between 1983 and 2024.}

\item{state}{A character string representing the state. Valid values are 'us' (contiguous U.S. and D.C.), 'hi' (Hawaii), and 'ak' (Alaska).}

\item{household_size}{An integer representing the household size. Valid values are between 1 and 8.}

\item{by_month}{A logical value indicating whether to return the poverty threshold as a monthly value. Default is FALSE.}
}
\value{
A tibble with columns 'year', 'state', 'household_size', and 'poverty_threshold'.
}
\description{
Fetches the federal poverty guideline for a given year, state, and household size from the HHS API.
The poverty guideline is the same for all states except Hawaii and Alaska.
}
