% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncpublicbenefits.R
\docType{package}
\name{ncpublicbenefits}
\alias{ncpublicbenefits}
\title{ncpublicbenefits: A Package for Calculating Public Benefits in NC}
\description{
This package provides functions to calculate public benefits such as Medicaid,
TANF, SNAP, housing vouchers, and more for North Carolina.
}
