% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plt_total_income}
\alias{plt_total_income}
\title{Plot Total Take-Home Income}
\usage{
plt_total_income(benefits_table, unique_composition, unique_benefits)
}
\arguments{
\item{benefits_table}{A tibble created by \code{create_benefits_table()}, containing household benefits data.}

\item{unique_composition}{A character vector specifying unique family compositions to include in the plot.}

\item{unique_benefits}{A character vector specifying unique benefits to include in the plot.}
}
\value{
A Highcharter plot object visualizing total take-home income by monthly wage.
}
\description{
Creates a line chart to visualize total take-home income by monthly wage income.
The chart compares "Income minus taxes" and "Income plus benefits, minus taxes."
}
